Integrate code block automation across chat entry points
